fix: relocate motion models
