+++ perception/multi_object_tracker/CMakeLists.txt
@@ -22,6 +22,12 @@ include_directories(
 # Generate exe file
 set(MULTI_OBJECT_TRACKER_SRC
   src/multi_object_tracker_core.cpp
+  src/data_association/data_association.cpp
+  src/data_association/mu_successive_shortest_path/mu_successive_shortest_path_wrapper.cpp
+  src/tracker/motion_model/motion_model_base.cpp
+  src/tracker/motion_model/bicycle_motion_model.cpp
+  src/tracker/motion_model/ctrv_motion_model.cpp
+  src/tracker/motion_model/cv_motion_model.cpp
   src/tracker/model/tracker_base.cpp
   src/tracker/model/big_vehicle_tracker.cpp
   src/tracker/model/normal_vehicle_tracker.cpp
@@ -31,12 +37,6 @@ set(MULTI_OBJECT_TRACKER_SRC
   src/tracker/model/pedestrian_and_bicycle_tracker.cpp
   src/tracker/model/unknown_tracker.cpp
   src/tracker/model/pass_through_tracker.cpp
-  src/data_association/data_association.cpp
-  src/data_association/mu_successive_shortest_path/mu_successive_shortest_path_wrapper.cpp
-  src/motion_model/motion_model_base.cpp
-  src/motion_model/bicycle_motion_model.cpp
-  src/motion_model/ctrv_motion_model.cpp
-  src/motion_model/cv_motion_model.cpp
 )
 
 ament_auto_add_library(multi_object_tracker_node SHARED
