+++ perception/multi_object_tracker/include/multi_object_tracker/tracker/model/bicycle_tracker.hpp
@@ -19,8 +19,8 @@
 #ifndef MULTI_OBJECT_TRACKER__TRACKER__MODEL__BICYCLE_TRACKER_HPP_
 #define MULTI_OBJECT_TRACKER__TRACKER__MODEL__BICYCLE_TRACKER_HPP_
 
-#include "multi_object_tracker/motion_model/bicycle_motion_model.hpp"
 #include "multi_object_tracker/tracker/model/tracker_base.hpp"
+#include "multi_object_tracker/tracker/motion_model/bicycle_motion_model.hpp"
 
 #include <kalman_filter/kalman_filter.hpp>
